Drop core module dependency workarounds intended for Red Hat
It should never have been upstream's job to worry about this.
inxi should just correctly expect all core modules to be present
and not defend against them being absent. The fact that some
Fedora/RHEL users were able to get into a situation where they
had inxi installed but not the required core modules was caused
by errors in the Fedora downstream inxi package.

I am fixing inxi on all current Fedora versions to ensure this
cannot happen with packaged inxi. Older EOL Fedora versions will
not get this change in their packaged inxi.

Anyone using inxi from a direct download on Fedora 28 or later,
or EL 8 or later, can be assumed to have all core modules present
if they did "dnf install perl". If they did not do "dnf install
perl" in order to use a Perl program, that is their problem, not
anybody else's (and it would be hard for the script to sensibly
defend against *all* errors that could possibly happen in that
case).

This also includes a bug fix: the sleep in set_sysctl_data tried
to use $sleep, which not declared in that context. As it was in
an `eval` block, this was hidden. Now it's not in an `eval`
block, this causes a compilation failure if not fixed.

Signed-off-by: Adam Williamson <awilliam@redhat.com>